Move configuration of HA to enviroment variables
This is done so that Kubernetes Secrets can be used, and secrets
don't have to be written to a file.

Also change the default in application.conf to ALFA and not ADMIN
so that only the first user gets to be ADMIN and not by accident
anyone who notices the sign up possibility.